Cleaned up to make it look more in order and was able to finish the delete function, as well as altered some functions to be compatible with the changes

+++ Menu.cpp
@@ -2,23 +2,36 @@
 
 void Menu::Main_Display()
 {
-    std::cout << "1.Add a person\n2.Make an account \n3.Delete a person \n4.Get person information \n5.Get account information \n6.List of people \n7.List of accounts " << std::endl;
+    std::cout << "1.Add a person\n2.Make an account \n3.Delete a person \n4.Get person information \n5.Get account information \n6.List of people \n7.List of accounts \n8.Quit" << std::endl;
 };
-void Menu::Add_Person_display()
+void Menu::Add_Person_display(People& people)
 {
     std::string name;
     std::cout << "Name to add: ";
     std::cin >> name;
     Person person(name);
+    people.addPerson(person);
 };
-void Menu::Delete_person_display(People people)
+void Menu::Delete_person_display(People& people, Accounts& accounts)
 {
     People_List_Display(people);
     std::cout << "Who would you like to delete, please input a valid name:" << std::endl;
     std::string find;
-    people.deletePerson(find);
+    int index;
+    bool exists = people.findPerson(find, index);
+    if(exists)
+    {
+        Account account = people.returnAtIndex(index).getAccount();
+        accounts.DeleteFromAccounts(account);
+        people.deletePerson(find);
+        std::cout << "Deletion was sucessfull." << std::endl;
+    }
+    else
+    {
+        std::cout << "Unable to delete as this person does not exist." << std::endl;
+    }
 };
-void Menu::Make_Account_Display(People people)
+void Menu::Make_Account_Display(People& people, Accounts& accounts)
 {
     std::string user_given_name;
     int index;
@@ -45,6 +58,7 @@ void Menu::Make_Account_Display(People people)
         people.returnAtIndex(index).getAccount().setAccountPassword(password);
         people.returnAtIndex(index).getAccount().setAccountUsername(username);
         people.returnAtIndex(index).setHasAccount();
+        accounts.AddToAccounts(people.returnAtIndex(index).getAccount());
     }
     else
     {
@@ -97,29 +111,24 @@ void Menu::Person_info_Display(People people, std::string name)
         std::cout << "The given name was not able to be found in the list. Unable to display a non-existant person's information." << std::endl; 
     }
 };
-void Menu::Person_account_Display(People people)
+void Menu::Person_account_Display(Accounts accounts)
 {
     std::string user_given_name;
-    People_List_Display(people);
+    accounts.print_AccountsList();
     std::cout << "Who would you want to see the person account information about?" << std::endl;
     std::cin >> user_given_name;
-    if(people.findPerson(user_given_name))
+    if(accounts.findPerson(user_given_name))
     {
-        Person person_to_display_account = people.returnAtName(user_given_name);
-        if(person_to_display_account.HaveAccount())
-        {
-            Account person_account = person_to_display_account.getAccount();
-            std::cout << "Name: " << person_to_display_account.getName() << std::endl;
-            std::cout << "Account Name: " << person_account.getAccountName() << std::endl;
-            std::cout << "Account Number: " << person_account.getAccountNum() << std::endl;
-            std::cout << "Account Username: " << person_account.getAccountUsername() << std::endl;
-            std::cout << "Account Password: " << person_account.getAccountPassword() << std::endl;
-            std::cout << "Account Four Digit Pin: " << person_account.get_four_digit_num() << std::endl; 
-        }
-        else
-        {
-            std::cout << "Unable to display the account since there is no account made for this person." << std::endl;
-        }
+        Account person_account = accounts.returnAtName(user_given_name);
+        std::cout << "Account Name: " << person_account.getAccountName() << std::endl;
+        std::cout << "Account Number: " << person_account.getAccountNum() << std::endl;
+        std::cout << "Account Username: " << person_account.getAccountUsername() << std::endl;
+        std::cout << "Account Password: " << person_account.getAccountPassword() << std::endl;
+        std::cout << "Account Four Digit Pin: " << person_account.get_four_digit_num() << std::endl; 
+    }
+    else
+    {
+        std::cout << "Unable to display the account since there is no account for a non-existant person." << std::endl;
     }
 };
 void Menu::People_List_Display(People people)
@@ -128,17 +137,10 @@ void Menu::People_List_Display(People people)
     people.PeopleList();
 
 };
-void Menu::People_with_account_Display(People people)
+void Menu::People_with_account_Display(Accounts accounts)
 {
     std::cout << "People with Account List: " << std::endl;
-    int number_of_people = people.numOfPeople();
-    for(int i = 0; i < number_of_people; i++)
-    {
-        if(people.returnAtIndex(i).HaveAccount())
-        {
-            Person_info_Display(people, people.returnAtIndex(i).getName());
-        }
-    }
+    accounts.print_AccountsList();
 };
